requirements.txt
IPy==0.82a
Pillow==2.6.1
miniupnpc==1.9
psutil==2.1.3
bitcoin==1.1.15
pycountry==1.9
pyelliptic==1.5.5
pysqlcipher==2.6.3-1
pystun==0.1.0
python-gnupg==0.3.6
python-obelisk==0.1.3
pyzmq==14.4.1
qrcode==5.1
requests==2.4.3
tornado==4.0.2
Twisted==14.0.2
rfc3986==0.2.0
dnschain==0.1.0
